requirements.txt
openai==2.14.0
redis>=7.1.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.0